requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
certifi==2024.12.14
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
motor==3.6.0
openai==1.59.7
passlib==1.7.4
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.10.4
pydantic_core==2.27.2
pymongo==4.9.2
python-decouple==3.8
python-jose==3.3.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
Telethon==1.38.1
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
